feat: Add tooltip to Community-LT button
