+++ src/components/TrackSelector/LiveTalkModalButton.tsx
@@ -1,6 +1,7 @@
 import React, { ReactElement, useState } from 'react'
 import * as Styled from './styled'
 import { FormatListBulleted } from '@material-ui/icons'
+import HtmlTooltip from './HtmlTooltip'
 
 type Props = {
   content: (closeModal: () => void) => ReactElement
@@ -11,13 +12,15 @@ export const LiveTalkModalButton: React.FC<Props> = ({ content }) => {
 
   return (
     <>
-      <Styled.LiveTalkModalButton
-        data-testid="btn"
-        color="primary"
-        onClick={() => setModalOpen(true)}
-      >
-        <FormatListBulleted />
-      </Styled.LiveTalkModalButton>
+      <HtmlTooltip title="ライブセッション">
+        <Styled.LiveTalkModalButton
+          data-testid="btn"
+          color="primary"
+          onClick={() => setModalOpen(true)}
+        >
+          <FormatListBulleted />
+        </Styled.LiveTalkModalButton>
+      </HtmlTooltip>
       <Styled.LiveTalkModal
         open={isModalOpen}
         onClose={() => setModalOpen(false)}
