+++ src/components/TrackSelector/PTrackSelectorButtonGroup.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 import * as Styled from './styled'
 import { Talk, Track } from '../../generated/dreamkast-api.generated'
-import { Theme, Tooltip } from '@material-ui/core'
 import dayjs from 'dayjs'
-import { withStyles } from '@material-ui/styles'
 import { setupDayjs } from '../../util/setupDayjs'
+import HtmlTooltip from './HtmlTooltip'
 
 setupDayjs()
 
@@ -61,13 +60,3 @@ export const PTrackSelectorButtonGroup: React.FC<Props> = ({
     </Styled.TrackSelectorButtonGroup>
   )
 }
-
-const HtmlTooltip = withStyles((theme: Theme) => ({
-  tooltip: {
-    backgroundColor: '#f5f5f9',
-    color: 'rgba(0, 0, 0, 0.87)',
-    maxWidth: 220,
-    fontSize: theme.typography.pxToRem(12),
-    border: '1px solid #dadde9',
-  },
-}))(Tooltip)
