+++ src/components/TrackSelector/ShowLTButton.tsx
@@ -3,12 +3,13 @@ import * as Styled from './styled'
 import { PeopleAlt } from '@material-ui/icons'
 import { showLT } from '../../store/settings'
 import { useDispatch } from 'react-redux'
+import HtmlTooltip from './HtmlTooltip'
 
 export const ShowLTButton: React.FC = () => {
   const dispatch = useDispatch()
 
   return (
-    <>
+    <HtmlTooltip title="コミュニティLT">
       <Styled.ShowLTButton
         data-testid="btn"
         color="primary"
@@ -18,6 +19,6 @@ export const ShowLTButton: React.FC = () => {
       >
         <PeopleAlt />
       </Styled.ShowLTButton>
-    </>
+    </HtmlTooltip>
   )
 }
